requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
dateparser==1.3.0
fonttools==4.62.1
frozenlist==1.8.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
pycryptodome==3.23.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.36
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
six==1.17.0
ta==0.11.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
yarl==1.23.0
